Refactor to generic RPC class.
