+++ TezosKit/Conseil/GetSentTransactionsRPC.swift
@@ -1,7 +1,7 @@
 // Copyright Keefer Taylor, 2019.
 
 /// An RPC which fetches sent transactions from an account.
-public class GetSentTransactionsRPC: RPC<[Transaction]> {
+public class GetSentTransactionsRPC: ConseilQueryRPC<[Transaction]> {
   /// - Parameters:
   ///   - account: The account to query.
   ///   - limit: The number of items to return.
@@ -9,32 +9,20 @@ public class GetSentTransactionsRPC: RPC<[Transaction]> {
   ///   - platform: The platform to query.
   ///   - network: The network to query.
   public init?(account: String, limit: Int, apiKey: String, platform: ConseilPlatform, network: ConseilNetwork) {
-    guard let escapedPlatform = platform.rawValue.addingPercentEncoding(
-                withAllowedCharacters: CharacterSet.urlQueryAllowed
-          ),
-          let escapedNetwork = network.rawValue.addingPercentEncoding(
-                withAllowedCharacters: CharacterSet.urlQueryAllowed
-      ) else {
-        return nil
-    }
-    let endpoint = "/v2/data/\(escapedPlatform)/\(escapedNetwork)/operations"
-    let headers = [
-      Header.contentTypeApplicationJSON,
-      Header(field: "apiKey", value: apiKey)
-    ]
-
     let predicates: [ConseilPredicate] = [
       ConseilQuery.Predicates.predicateWith(field: "kind", set: ["transaction"]),
       ConseilQuery.Predicates.predicateWith(field: "source", set: [account])
     ]
     let orderBy: ConseilOrderBy = ConseilQuery.OrderBy.orderBy(field: "timestamp")
-    let payload: [String: Any] = ConseilQuery.query(predicates: predicates, orderBy: orderBy, limit: limit)
+    let query: [String: Any] = ConseilQuery.query(predicates: predicates, orderBy: orderBy, limit: limit)
 
     super.init(
-      endpoint: endpoint,
-      headers: headers,
-      responseAdapterClass: TransactionsResponseAdapter.self,
-      payload: JSONUtils.jsonString(for: payload)
+      query: query,
+      path: "operations",
+      apiKey: apiKey,
+      platform: platform,
+      network: network,
+      responseAdapterClass: TransactionsResponseAdapter.self
     )
   }
 }
